ユークリッド距離の計算メソッドを追加
