+++ NavigationForiOS/Lpf.swift
@@ -21,9 +21,9 @@ class Lpf: AlgorithmBase {
     ///
     /// - Parameters:
     ///   - navigations: ナビゲーションのルートなどの情報を含む変数
-    ///   - receivedBeaconsRssi: 実際に歩いて取得できたビーコンのRSSIの値（平滑化済み）
+    ///   - receivedBeaconsRssi: 現在のビーコンのRSSIの値（平滑化済み）
     /// - Returns: return 現在の場所のENUM
-    func getCurrentPoint(navigations: NavigationEntity, receivedBeaconsRssi : Dictionary<Int, Int>) -> POINT {
+    override func getCurrentPoint(navigations: NavigationEntity, receivedBeaconsRssi : Dictionary<Int, Int>) -> POINT {
         let maxRssiMinorId = self.getMaxRssiMinorId(receivedBeaconsRssi: receivedBeaconsRssi)
         let threshold = navigations.getBeaconThresholdFromMinorId(minor_id: maxRssiMinorId)
         if(self.isOnNavigationPoint(RSSI: receivedBeaconsRssi[maxRssiMinorId]!, threshold: threshold)){
@@ -35,35 +35,6 @@ class Lpf: AlgorithmBase {
             return POINT.CROSSROAD
         }
         return POINT.OTHER
-    }
-    
-    
-    /// 最大値のRSSIを取得する関数
-    ///
-    /// - Parameter receivedBeaconsRssi: ナビゲーションで利用しているビーコンのRssiの値(平滑化済み)
-    /// - Returns: 最大のRSSIの値を持つminorId
-    func getMaxRssiMinorId(receivedBeaconsRssi: Dictionary<Int, Int>) -> Int {
-        var maxMinorId = -1
-        var maxRssi = -1000
-        receivedBeaconsRssi.forEach { (key: Int, value: Int) in
-            if(maxRssi < value) {
-                maxMinorId = key
-                maxRssi = value
-            }
-        }
-        return maxMinorId
-    }
-    
-    
-    /// 現在のルートIDを取得する関数
-    ///
-    /// - Parameters:
-    ///   - navigations: ナビゲーションのルートなどの情報を含む変数
-    ///   - receivedBeaconsRssi: 実際に歩いて取得できたビーコンのRSSIの値（平滑化済み）
-    /// - Returns: 現在のルートID
-    func getRouteId(navigations: NavigationEntity, receivedBeaconsRssi: Dictionary<Int, Int>) -> Int {
-        var target = self.getMaxRssiMinorId(receivedBeaconsRssi: receivedBeaconsRssi)
-        return navigations.getRouteIdFromMinorId(minor_id: target)
     }
     
     //ナビゲーションを行うタイミングを判定する
